+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## Unreleased
+
+SECURITY:
+
+FEATURES:
+
+IMPROVEMENTS:
+
+BUG FIXES:
+
 ## v0.3.0 (August 8, 2019)
 
 Refactor. EVM-Lite becomes a library